+++ src/constants.ts
@@ -1,4 +1,4 @@
-export const VERSION = "3.2.0";
+export const VERSION = "3.3.0";
 
 export const Status = {
   QUEUED: "queued",